renamed and arranged POM files
