+++ cypress/POM/loginPage.js
@@ -2,9 +2,11 @@ class LoginPage {
     get emailInput() {
         return cy.get("#email");
     }
+
     get passwordInput() {
         return cy.get("#password");
     }
+
     get submitButton() {
         return cy.get("button[type='submit']");
     }
