+++ cypress/POM/viewAdPage.js
